requirements.txt
langchain
langchain-openai
langchain-community
openai
chromadb
streamlit
pandas
requests
python-dotenv
tiktoken
